+++ git-repository/src/easy/ext/mod.rs
@@ -1,0 +1,6 @@
+#![allow(missing_docs)]
+pub(crate) mod object;
+pub use object::ObjectAccessExt;
+
+mod reference;
+pub use reference::ReferenceAccessExt;
