+++ git-repository/src/easy/oid.rs
@@ -58,12 +58,12 @@ where
 {
     // NOTE: Can't access other object data that is attached to the same cache.
     pub fn object(&self) -> Result<ObjectRef<'repo, A>, find::existing::Error> {
-        crate::ext::access::object::find_object(self.access, self.id)
+        crate::easy::ext::object::find_object(self.access, self.id)
     }
 
     // NOTE: Can't access other object data that is attached to the same cache.
     pub fn try_object(&self) -> Result<Option<ObjectRef<'repo, A>>, find::Error> {
-        crate::ext::access::object::try_find_object(self.access, self.id)
+        crate::easy::ext::object::try_find_object(self.access, self.id)
     }
 }
