+++ git-repository/src/ext/mod.rs
@@ -1,8 +1,5 @@
-mod tree;
+pub use object_id::ObjectIdExt;
 pub use tree::TreeIterExt;
 
 mod object_id;
-pub use object_id::ObjectIdExt;
-
-pub(crate) mod access;
-pub use access::{object::ObjectAccessExt, reference::ReferenceAccessExt};
+mod tree;
